+++ geocoding/test/geocoding_test.dart
@@ -4,6 +4,8 @@ import 'package:mockito/mockito.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 final mockLocation = Location(
+  title: '',
+  description: '',
   latitude: 52.2165157,
   longitude: 6.9437819,
   timestamp: DateTime.fromMillisecondsSinceEpoch(0).toUtc(),
@@ -48,8 +50,9 @@ class MockGeocodingPlatform extends Mock
         GeocodingPlatform {
   @override
   Future<List<Location>> locationFromAddress(
-    String address,
-  ) async {
+    String address, {
+    Region? targetRegion,
+  }) async {
     return [mockLocation];
   }