+++ geocoding_platform_interface/lib/src/models/location.dart
@@ -1,4 +1,4 @@
-import 'package:meta/meta.dart';
+import 'package:flutter/foundation.dart';
 
 /// Contains detailed location information.
 @immutable
@@ -7,16 +7,26 @@ class Location {
   /// instances constructed this way won't actually reflect any real information
   /// from the platform, just whatever was passed in at construction time.
   const Location({
+    required this.title,
+    required this.description,
     required this.latitude,
     required this.longitude,
     required this.timestamp,
   });
 
   const Location._({
+    required this.title,
+    required this.description,
     required this.latitude,
     required this.longitude,
     required this.timestamp,
   });
+
+  /// The title associated with the placemark, might be empty
+  final String title;
+
+  /// The description associated with the placemark, might be empty
+  final String description;
 
   /// The latitude associated with the placemark.
   final double latitude;
@@ -30,6 +40,8 @@ class Location {
   @override
   bool operator ==(Object other) =>
       other is Location &&
+      other.title == title &&
+      other.description == description &&
       other.latitude == latitude &&
       other.longitude == longitude &&
       other.timestamp == timestamp;
@@ -65,6 +77,8 @@ class Location {
     }
 
     return Location._(
+      title: locationMap['title'] ?? '',
+      description: locationMap['description'] ?? '',
       latitude: locationMap['latitude'],
       longitude: locationMap['longitude'],
       timestamp: timestamp,
@@ -74,6 +88,8 @@ class Location {
   /// Converts the [Location] instance into a [Map] instance that can be
   /// serialized to JSON.
   Map<String, dynamic> toJson() => {
+        'title': title,
+        'description': description,
         'latitude': latitude,
         'longitude': longitude,
         'timestamp': timestamp.millisecondsSinceEpoch,
@@ -82,6 +98,8 @@ class Location {
   @override
   String toString() {
     return '''
+      Title: $title,
+      Description: $description,
       Latitude: $latitude,
       Longitude: $longitude,
       Timestamp: $timestamp''';