+++ geocoding_platform_interface/lib/src/models/models.dart
@@ -1,2 +1,3 @@
 export 'location.dart';
 export 'placemark.dart';
+export 'region.dart';